requirements.txt
mediapipe==0.10.21
opencv_contrib_python==4.11.0.86
pygame==2.6.1
pywebview==6.1
